renamed "EndpointResolverTests.cs" to "RequestCoordinatorTests.cs" as that is now the name of the class under test
