+++ src/SevenDigital.Api.Wrapper.Unit.Tests/SevenDigital.Api.Wrapper.Unit.Tests.csproj
@@ -62,7 +62,7 @@
     <Compile Include="Deserialisation\User\Payment\Cards_unit_tests.cs" />
     <Compile Include="Deserialisation\User\Payment\AddCardsTests.cs" />
     <Compile Include="Deserialisation\User\Payment\DeleteCardTests.cs" />
-    <Compile Include="EndpointResolution\EndpointResolverTests.cs" />
+    <Compile Include="EndpointResolution\RequestCoordinatorTests.cs" />
     <Compile Include="EndpointResolution\OAuth\NonceFactoryConcurrenyTests.cs" />
     <Compile Include="EndpointResolution\OAuth\OAuthSignatureGeneratorTests.cs" />
     <Compile Include="Exceptions\ApiWebExceptionTests.cs" />
